Cleanup duplicate auth code
